refactor: consolidate media helpers
